update DB and update score to float

+++ hustack-backend/baseweb/src/main/java/com/hust/baseweb/applications/exam/model/request/ExamMarkingSaveReq.java
@@ -13,7 +13,7 @@ public class ExamMarkingSaveReq {
 
     private String examResultId;
     private String comment;
-    private Integer totalScore;
+    private Float totalScore;
     private List<String> commentFilePathDeletes;
     private List<ExamMarkingDetailsSaveReq> examResultDetails;
 }
