+++ hustack-backend/baseweb/src/main/java/com/hust/baseweb/applications/exam/model/response/ExamMarkingDetailsResDB.java
@@ -9,7 +9,7 @@ public interface ExamMarkingDetailsResDB {
     String getExamStudentEmail();
     String getExamStudentPhone();
     String getExamResultId();
-    Integer getTotalScore();
+    Float getTotalScore();
     Integer getTotalTime();
     String getSubmitedAt();
     String getAnswerFiles();
