+++ hustack-backend/baseweb/src/main/java/com/hust/baseweb/applications/exam/model/response/MyExamQuestionDetailsRes.java
@@ -19,7 +19,7 @@ public interface MyExamQuestionDetailsRes {
     String getAnswer();
     String getFilePathAnswer();
     String getFilePathComment();
-    Integer getScore();
+    Float getScore();
     @JsonRawValue
     String getQuestionAnswers();
 }
